python/n1swmsapp/requirements.txt
Django==3.1.1
Pillow
djangorestframework
openpyxl
mysql
django-truncate
